Splitting abstractions into different files
