+++ adapters/repos/db/vector/compressionhelpers/product_quantization.go
@@ -327,7 +327,7 @@ func (pq *ProductQuantizer) NewDistancer(a []float32) *PQDistancer {
 	}
 }
 
-func (pq *ProductQuantizer) NewCompressedQuantizerDistancer(a []byte) QuantizerDistancer[byte] {
+func (pq *ProductQuantizer) NewCompressedQuantizerDistancer(a []byte) quantizerDistancer[byte] {
 	return &PQDistancer{
 		x:          nil,
 		pq:         pq,
